refactor: 💡 stylelint and packers

+++ src/commands/build/build.command.class.ts
@@ -1,15 +1,79 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 /* eslint-disable new-cap */
 /* eslint-disable @typescript-eslint/no-this-alias */
+const SYNC_WRITE_METHODS = [
+    'writeFile',
+    'writeFileSync',
+    'appendFile',
+    'appendFileSync',
+    'mkdir',
+    'mkdirSync',
+    'rmdir',
+    'rmdirSync',
+    'unlink',
+    'unlinkSync',
+];
+const PROMISE_WRITE_METHODS = ['writeFile', 'appendFile', 'mkdir', 'rmdir', 'unlink'];
+let interceptWriting = false;
+
+import { Volume } from 'memfs';
+import { ufs } from 'unionfs';
+import { Module } from 'module';
+import * as fs from 'fs-extra';
 import * as webpack from 'webpack';
-import { resolve as pathResolve } from 'path';
+import { dirname as pathDirname, resolve as pathResolve } from 'path';
 import { Command } from 'commander';
 import * as yup from 'yup';
-import * as fs from 'fs-extra';
 import * as childProcess from 'child_process';
 import { StringUtil } from '../../utilities/string-util.class';
 import * as _ from 'lodash';
 import { CommandFactory } from '../command-factory.class';
 import * as winston from 'winston';
+
+const volume = new Volume();
+
+ufs.use(fs).use(volume as unknown as typeof fs);
+
+// @ts-ignore
+const originalModuleLoad = Module._load;
+
+// @ts-ignore
+Module._load = (request, parent, isMain) => {
+    const promiseFs = {
+        ...ufs.promises,
+        ...PROMISE_WRITE_METHODS.reduce((result, methodName) => {
+            result[methodName] = interceptWriting
+                ? async (...args) => {
+                      return volume.promises[methodName](...args);
+                  }
+                : ufs.promises[methodName];
+            return result;
+        }, {}),
+    };
+
+    if (request === 'fs' || request === 'node:fs') {
+        return {
+            ...ufs,
+            ...SYNC_WRITE_METHODS.reduce((result, methodName) => {
+                result[methodName] = interceptWriting
+                    ? (...args) => {
+                          return volume[methodName](...args);
+                      }
+                    : ufs[methodName];
+                return result;
+            }, {}),
+            promises: promiseFs,
+        };
+    }
+
+    if (request === 'fs/promises' || request === 'node:fs/promises') {
+        return promiseFs;
+    }
+
+    return originalModuleLoad(request, parent, isMain);
+};
+
+// import { exec } from '@yao-pkg/pkg';
 
 class CatchNotFoundPlugin {
     public constructor(private logger?: winston.Logger) {}
@@ -23,19 +87,29 @@ class CatchNotFoundPlugin {
                 const notfoundPathname = pathResolve(__dirname, '../../../') + `/preserved/@@notfound.js?${request}`;
                 if (result) {
                     return callback(null, innerPath, result);
-                };
-                if (error && !error.message.startsWith('Can\'t resolve')) {
+                }
+                if (error && !error.message.startsWith("Can't resolve")) {
                     return callback(error);
                 }
                 // Allow .js resolutions to .tsx? from .tsx?
-                if (request.endsWith('.js') && context.issuer && (context.issuer.endsWith('.ts') || context.issuer.endsWith('.tsx'))) {
-                    return resolve.call(self, context, path, request.slice(0, -3), resolveContext, (error1, innerPath, result) => {
-                        if (result) return callback(null, innerPath, result);
-                        if (error1 && !error1.message.startsWith('Can\'t resolve'))
-                            return callback(error1);
-                        // make not found errors runtime errors
-                        callback(null, notfoundPathname, request);
-                    });
+                if (
+                    request.endsWith('.js') &&
+                    context.issuer &&
+                    (context.issuer.endsWith('.ts') || context.issuer.endsWith('.tsx'))
+                ) {
+                    return resolve.call(
+                        self,
+                        context,
+                        path,
+                        request.slice(0, -3),
+                        resolveContext,
+                        (error1, innerPath, result) => {
+                            if (result) return callback(null, innerPath, result);
+                            if (error1 && !error1.message.startsWith("Can't resolve")) return callback(error1);
+                            // make not found errors runtime errors
+                            callback(null, notfoundPathname, request);
+                        },
+                    );
                 }
                 logger?.warn?.(`Notfound '${context.issuer}' from '${request}', skipping...`);
                 // make not found errors runtime errors
@@ -105,21 +179,34 @@ class AutoRunPlugin {
 }
 
 class CleanPlugin {
-    public apply(compiler) {
+    public constructor(private readonly outputJsPathname: string) {}
+
+    public apply(compiler: webpack.Compiler) {
         compiler.hooks.compilation.tap('CleanPlugin', (compilation) => {
-            compilation.hooks.processAssets.tap(
-                {
-                    name: 'CleanPlugin',
-                    stage: compilation.PROCESS_ASSETS_STAGE_OPTIMIZE,
-                },
-                (assets) => {
-                    Object.keys(assets).forEach((key) => {
-                        if (!key?.endsWith?.('.js')) {
-                            _.unset(assets, key);
-                        }
-                    });
-                },
-            );
+            compilation.hooks.processAssets.tap('CleanPlugin', (assets) => {
+                Object.keys(assets).forEach((key) => {
+                    if (pathResolve(pathDirname(this.outputJsPathname), key) !== this.outputJsPathname) {
+                        _.unset(assets, key);
+                    }
+                });
+            });
+        });
+    }
+}
+
+class BinaryPlugin {
+    public constructor(private readonly outputJsPathname: string) {}
+
+    public apply(compiler: webpack.Compiler) {
+        compiler.hooks.compilation.tap('BinaryPlugin', (compilation) => {
+            compilation.hooks.processAssets.tapPromise('BinaryPlugin', async (assets) => {
+                const asset = Object.entries(assets).find(
+                    ([relativePathname]) =>
+                        pathResolve(pathDirname(this.outputJsPathname), relativePathname) === this.outputJsPathname,
+                )?.[1];
+                // volume.writeFileSync(this.outputJsPathname, asset?.source());
+                console.log('LENCONDA', asset?.source?.());
+            });
         });
     }
 }
@@ -140,10 +227,7 @@ export const BuildCommand = CommandFactory.create({
     } as {
         childProcess: childProcess.ChildProcess;
     },
-    register: ({
-        command,
-        callback,
-    }) => {
+    register: ({ command, callback }) => {
         command.addCommand(
             new Command('build')
                 .option('--clean', 'Clean output directory')
@@ -156,16 +240,11 @@ export const BuildCommand = CommandFactory.create({
                 .action(callback),
         );
     },
-    run: ({
-        logger,
-        options,
-        context,
-    }) => {
-        const {
-            watch,
-            clean,
-            ...webpackOptions
-        } = options;
+    run: ({ logger, options, context }) => {
+        const { watch, clean, ...webpackOptions } = options;
+        interceptWriting = !watch;
+        const outputDirectoryPathname = pathResolve(webpackOptions.workDir, webpackOptions.outputPath);
+        const outputFilePathname = pathResolve(outputDirectoryPathname, `${webpackOptions.name}.js`);
         const compiler = webpack({
             optimization: {
                 minimize: false,
@@ -174,10 +253,10 @@ export const BuildCommand = CommandFactory.create({
                 [webpackOptions.name]: pathResolve(webpackOptions.workDir, webpackOptions.entry),
             },
             target: 'node',
-            mode: options.watch ? 'development' : 'production',
+            mode: watch ? 'development' : 'production',
             output: {
                 filename: webpackOptions.outputFilename,
-                path: pathResolve(webpackOptions.workDir, webpackOptions.outputPath),
+                path: outputDirectoryPathname,
                 libraryTarget: 'commonjs',
             },
             resolve: {
@@ -186,9 +265,7 @@ export const BuildCommand = CommandFactory.create({
                     src: pathResolve(webpackOptions.workDir, './src'),
                     UNKNOWN: false,
                 },
-                plugins: [
-                    new CatchNotFoundPlugin(logger),
-                ],
+                plugins: [new CatchNotFoundPlugin(logger)],
             },
             module: {
                 rules: [
@@ -209,21 +286,27 @@ export const BuildCommand = CommandFactory.create({
                 new webpack.ProgressPlugin((percentage, message) => {
                     logger?.info?.(`Build progress: ${percentage * 100}%, ${message}`);
                 }),
-                new CleanPlugin(),
-                ...(!watch ? [] : [
-                    new AutoRunPlugin({
-                        logger,
-                        parallel: true,
-                        onAfterStart: (childProcess) => {
-                            context.childProcess = childProcess;
-                        },
-                        onBeforeStart: () => {
-                            if (context.childProcess) {
-                                context.childProcess.kill();
-                            }
-                        },
-                    }),
-                ]),
+                new CleanPlugin(outputFilePathname),
+                ...(() => {
+                    if (watch) {
+                        return [
+                            new AutoRunPlugin({
+                                logger,
+                                parallel: true,
+                                onAfterStart: (childProcess) => {
+                                    context.childProcess = childProcess;
+                                },
+                                onBeforeStart: () => {
+                                    if (context.childProcess) {
+                                        context.childProcess.kill();
+                                    }
+                                },
+                            }),
+                        ];
+                    } else {
+                        return [new BinaryPlugin(outputFilePathname)];
+                    }
+                })(),
             ],
         });
